Added updates to support viewing and editting the answer for a given option.
Made sure that on devices, we attempt to move the UI up when the keyboard arrives so that the user can see what they're doing.
Strangely, I can test on the simulator, but not on a real device. Conversely, I can save the data between invocations on a real device but not in the simulator.